+++ JayYi/ArchitectureStudy/app/src/main/java/com/jay/architecturestudy/ui/movie/MovieAdapter.kt
@@ -1,4 +1,4 @@
-package com.jay.architecturestudy.ui
+package com.jay.architecturestudy.ui.movie
 
 import android.content.Context
 import android.content.Intent
@@ -6,12 +6,12 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
 import androidx.core.text.HtmlCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.jay.architecturestudy.R
 import com.jay.architecturestudy.model.Movie
+import com.jay.architecturestudy.ui.WebViewActivity
 import com.jay.architecturestudy.ui.WebViewActivity.Companion.EXTRA_URL
 import kotlinx.android.synthetic.main.list_item_movie.view.*
 import java.lang.Exception
