+++ JayYi/ArchitectureStudy/app/src/main/java/com/jay/architecturestudy/ui/movie/MovieFragment.kt
@@ -1,4 +1,4 @@
-package com.jay.architecturestudy.ui
+package com.jay.architecturestudy.ui.movie
 
 import android.os.Bundle
 import android.util.Log
@@ -10,7 +10,7 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.jay.architecturestudy.R
-import com.jay.architecturestudy.model.ResponseKin
+import com.jay.architecturestudy.model.ResponseMovies
 import com.jay.architecturestudy.network.Api
 import kotlinx.android.synthetic.main.fragemnt_movie.*
 import kotlinx.android.synthetic.main.view_search.*
@@ -18,15 +18,17 @@ import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
-class KinFragment : Fragment() {
-    private lateinit var kinAdapter: KinAdapter
+
+class MovieFragment : Fragment() {
+
+    private lateinit var movieAdapter: MovieAdapter
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragemnt_kin, container, false)
+        return inflater.inflate(R.layout.fragemnt_movie, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -39,9 +41,8 @@ class KinFragment : Fragment() {
                 search(keyword)
             }
         }
-
         activity?.let { activity ->
-            kinAdapter = KinAdapter(activity)
+            movieAdapter = MovieAdapter(activity)
                 .also {
                     recycler_view.apply {
                         adapter = it
@@ -58,19 +59,19 @@ class KinFragment : Fragment() {
     }
 
     private fun search(keyword: String) {
-        Api.getKin(keyword)
-            .enqueue(object : Callback<ResponseKin> {
-                override fun onFailure(call: Call<ResponseKin>, t: Throwable) {
-                    Log.e("Kin", "error=${t.message}")
+        Api.getMovies(keyword)
+            .enqueue(object : Callback<ResponseMovies> {
+                override fun onFailure(call: Call<ResponseMovies>, t: Throwable) {
+                    Log.e("Movie", "error=${t.message}")
                 }
 
                 override fun onResponse(
-                    call: Call<ResponseKin>,
-                    response: Response<ResponseKin>
+                    call: Call<ResponseMovies>,
+                    response: Response<ResponseMovies>
                 ) {
                     if (response.isSuccessful) {
                         val body = response.body() ?: return
-                        kinAdapter.setData(body.kins)
+                        movieAdapter.setData(body.movies)
                     }
                 }
